dublicate code moved to -> templates.py
