+++ brain_games/task6_code.py
@@ -1,9 +1,7 @@
 from random import randint, choice
 
-import prompt
+from .templates import get_resolution, get_user_value
 
-MSG_WRONG_ANSWER = "is wrong answer ;(. Correct answer was"
-MSG_RIGHT_ANSWER = "Correct!"
 ATTEMPS = 3
 MIN_NUM = 1
 MAX_NUM = 10
@@ -40,27 +38,6 @@ def get_result(question: list) -> int:
 
     elif action == "*":
         return a * b
-
-
-def get_user_value() -> int:
-
-    return prompt.integer("Your answer: ")
-
-
-def get_resolution(result: int, user_value: int) -> bool:
-    """
-    Function checks the result of the user
-    :param result: right result
-    :param user_value: user input
-    :return: resolution
-    """
-    if result == user_value:
-        print(MSG_RIGHT_ANSWER)
-        return True
-
-    else:
-        print(f"{user_value} {MSG_WRONG_ANSWER} {result},")
-        return False
 
 
 def start_game():
